storage: extract CompactedEvent to storage_types
